FBP-59. Extracted test events
