+++ flowable-xml-parser/src/test/kotlin/com/valb3r/bpmn/intellij/plugin/flowable/parser/FlowableParserDumbTest.kt
@@ -8,18 +8,16 @@ import com.valb3r.bpmn.intellij.plugin.bpmn.api.bpmn.elements.gateways.BpmnExclu
 import com.valb3r.bpmn.intellij.plugin.bpmn.api.diagram.DiagramElementId
 import com.valb3r.bpmn.intellij.plugin.bpmn.api.diagram.elements.BoundsElement
 import com.valb3r.bpmn.intellij.plugin.bpmn.api.diagram.elements.ShapeElement
-import com.valb3r.bpmn.intellij.plugin.bpmn.api.diagram.elements.WaypointElement
-import com.valb3r.bpmn.intellij.plugin.bpmn.api.events.*
 import com.valb3r.bpmn.intellij.plugin.bpmn.api.info.Property
 import com.valb3r.bpmn.intellij.plugin.bpmn.api.info.PropertyType
+import com.valb3r.bpmn.intellij.plugin.flowable.parser.testevents.*
 import org.amshove.kluent.*
 import org.junit.jupiter.api.Test
-import java.awt.geom.Point2D
 import java.nio.charset.StandardCharsets.UTF_8
 import java.util.*
 
 
-internal class DumbFlowableParserTest {
+internal class FlowableParserDumbTest {
 
     private val parentElemId = BpmnElementId("duplicates")
     private val bmpnElemId = BpmnElementId(UUID.randomUUID().toString())
@@ -286,70 +284,3 @@ internal class DumbFlowableParserTest {
     fun String.asResource(): String? = object {}::class.java.classLoader.getResource(this)?.readText(UTF_8)
 }
 
-data class WaypointElementState (
-        override val id: DiagramElementId,
-        override val x: Float,
-        override val y: Float,
-        override val origX: Float,
-        override val origY: Float,
-        override val physical: Boolean,
-        override val internalPhysicalPos: Int
-): IdentifiableWaypoint {
-
-    override fun moveTo(dx: Float, dy: Float): IdentifiableWaypoint {
-        TODO("Not yet implemented")
-    }
-
-    override fun asPhysical(): IdentifiableWaypoint {
-        TODO("Not yet implemented")
-    }
-
-    override fun originalLocation(): IdentifiableWaypoint {
-        TODO("Not yet implemented")
-    }
-
-    override fun asWaypointElement(): WaypointElement {
-        TODO("Not yet implemented")
-    }
-
-    override fun copyAndTranslate(dx: Float, dy: Float): IdentifiableWaypoint {
-        TODO("Not yet implemented")
-    }
-}
-
-data class EdgeElementState  (
-        override val id: DiagramElementId,
-        override val bpmnElement: BpmnElementId?,
-        override val waypoint: MutableList<IdentifiableWaypoint> = mutableListOf(),
-        override val epoch: Int
-): EdgeWithIdentifiableWaypoints {
-
-    override fun updateBpmnElemId(newId: BpmnElementId): EdgeWithIdentifiableWaypoints {
-        TODO("Not yet implemented")
-    }
-}
-
-data class StringValueUpdatedEvent(override val bpmnElementId: BpmnElementId, override val property: PropertyType, override val newValue: String, override val referencedValue: String? = null, override val newIdValue: BpmnElementId? = null): PropertyUpdateWithId
-
-data class BooleanValueUpdatedEvent(override val bpmnElementId: BpmnElementId, override val property: PropertyType, override val newValue: Boolean, override val referencedValue: Boolean? = null, override val newIdValue: BpmnElementId? = null): PropertyUpdateWithId
-
-data class DraggedToEvent(override val diagramElementId: DiagramElementId, override val dx: Float, override val dy: Float, override val parentElementId: DiagramElementId?, override val internalPos: Int?): LocationUpdateWithId
-
-data class NewWaypointsEvent(override val edgeElementId: DiagramElementId, override val waypoints: List<IdentifiableWaypoint>, override val epoch: Int): NewWaypoints
-
-data class DiagramElementRemovedEvent(override val elementId: DiagramElementId): DiagramElementRemoved
-
-data class BpmnElementRemovedEvent(override val elementId: BpmnElementId): BpmnElementRemoved
-
-data class BpmnShapeObjectAddedEvent(override val bpmnObject: WithParentId, override val shape: ShapeElement, override val props: Map<PropertyType, Property>): BpmnShapeObjectAdded
-
-data class BpmnEdgeObjectAddedEvent(override val bpmnObject: WithParentId, override val edge: EdgeWithIdentifiableWaypoints, override val props: Map<PropertyType, Property>): BpmnEdgeObjectAdded
-
-data class BpmnShapeResizedAndMovedEvent(override val diagramElementId: DiagramElementId, override val cx: Float, override val cy: Float, override val coefW: Float, override val coefH: Float): BpmnShapeResizedAndMoved {
-
-    override fun transform(point: Point2D.Float): Point2D.Float {
-        return Point2D.Float(cx + (point.x - cx) * coefW, cy + (point.y - cy) * coefH)
-    }
-}
-
-data class BpmnParentChangedEvent(override val bpmnElementId: BpmnElementId, override val newParentId: BpmnElementId, override val propagateToXml: Boolean = true): BpmnParentChanged
